fix(clippy): resolve warnings and apply code formatting
- Remove redundant closures in agent explorer module
- Remove unnecessary type cast in zip archive provider
- Apply cargo fmt formatting across workspace
- Simplify entry_stream implementation

+++ backend/opsbox-core/src/odfs/providers/archive/zip.rs
@@ -56,7 +56,7 @@ impl OpsFileSystem for ZipOpsFS {
         } else {
           OpsFileType::File
         },
-        size: entry.uncompressed_size() as u64,
+        size: entry.uncompressed_size(),
         modified: None,
         mode: 0,
         mime_type: None,
